City of St. Louis Boundary

A simple features data set containing a single polygon with the extent of the City of St. Louis's boundary.

data(stlBoundary)

Format

A simple feature with 1 rows and 6 variables:

LAYER

polygon description

AREA_SQMI

square mileage of city

Name

polygon name

Shape_Leng

length of polygon

Shape_Area

area of polygon

geometry

simple features geometry

Examples

class(stlBoundary)

#> [1] "sf"         "data.frame"

str(stlBoundary)

#> Classes ‘sf’ and 'data.frame':	1 obs. of  6 variables:
#>  $ LAYER     : Factor w/ 1 level "--CITY_LIMIT_BOUND_2002--": 1
#>  $ AREA_SQMI : num 62.1
#>  $ Name      : Factor w/ 1 level "St. Louis": 1
#>  $ Shape_Leng: num 218276
#>  $ Shape_Area: num 1.85e+09
#>  $ geometry  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:113, 1:2] 891596 889496 889494 888680 887971 ...
#>   .. ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>   ..- attr(*, "n_empty")= int 0
#>   ..- attr(*, "crs")=List of 2
#>   .. ..$ epsg       : int NA
#>   .. ..$ proj4string: chr "+proj=tmerc +lat_0=35.83333333333334 +lon_0=-90.5 +k=0.9999333333333333 +x_0=250000 +y_0=0 +datum=NAD83 +units=us-ft +no_defs"
#>   .. ..- attr(*, "class")= chr "crs"
#>   ..- attr(*, "class")= chr  "sfc_POLYGON" "sfc"
#>   ..- attr(*, "precision")= num 0
#>   ..- attr(*, "bbox")=Class 'bbox'  Named num [1:4] 871507 982635 915344 1070958
#>   .. .. ..- attr(*, "names")= chr [1:4] "xmin" "ymin" "xmax" "ymax"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   ..- attr(*, "names")= chr  "LAYER" "AREA_SQMI" "Name" "Shape_Leng" ...

head(stlBoundary)

#>                       LAYER AREA_SQMI      Name Shape_Leng Shape_Area
#> 1 --CITY_LIMIT_BOUND_2002--     62.11 St. Louis   218275.8 1845207811
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                geometry
#> 1 891596.1, 889496.1, 889494.0, 888679.6, 887971.3, 887786.8, 887590.3, 887304.4, 886840.2, 886519.4, 886418.9, 886258.7, 885899.0, 885695.6, 885595.0, 885596.8, 885842.0, 885497.5, 885535.5, 885473.8, 885360.4, 885339.4, 885325.1, 885314.2, 884811.3, 883718.8, 883398.1, 883312.4, 882017.9, 881062.1, 880322.7, 879570.6, 878938.1, 878268.0, 878222.5, 873493.6, 872737.2, 872711.7, 871506.7, 874234.0, 874198.3, 874853.3, 875104.6, 875494.2, 875956.4, 876298.2, 876799.4, 891036.2, 897787.4, 898855.3, 899577.4, 900143.0, 900548.7, 900654.2, 902172.4, 903507.5, 904592.3, 905200.9, 906527.3, 907056.5, 908042.9, 908817.7, 909084.8, 909158.4, 909661.2, 910005.4, 910181.8, 910337.8, 910241.0, 912829.1, 915302.4, 915343.8, 915313.6, 913924.6, 912849.3, 912514.6, 910499.5, 908251.2, 904299.1, 902965.4, 902542.9, 902404.1, 902382.7, 903192.0, 905200.7, 907205.7, 908687.3, 909733.5, 909970.5, 910830.2, 911265.8, 912258.1, 912245.7, 912184.0, 911961.3, 910982.5, 910355.0, 910217.7, 908284.3, 907089.4, 905644.0, 905238.5, 903071.6, 901190.6, 900289.0, 899649.5, 899356.2, 899050.6, 895109.6, 893551.8, 892019.2, 891712.7, 891596.1, 986509.4, 982638.9, 982634.9, 983053.9, 983380.6, 983296.1, 983252.6, 983248.1, 983460.5, 984016.9, 984435.9, 984644.9, 984884.3, 985051.5, 985280.7, 985581.4, 986134.4, 987195.7, 987301.2, 987631.3, 988002.4, 988506.5, 988568.9, 988595.9, 988946.8, 989369.0, 989649.6, 989726.3, 990609.3, 991188.3, 991736.4, 992432.7, 993013.4, 993638.7, 993691.7, 999625.6, 1001762.5, 1001834.2, 1005229.4, 1014023.5, 1014053.5, 1018372.2, 1019182.5, 1019557.3, 1022827.7, 1023741.2, 1027642.6, 1050692.1, 1054098.4, 1054455.5, 1054819.9, 1055615.3, 1056012.4, 1056074.4, 1057053.4, 1058325.0, 1059589.6, 1060286.4, 1061822.6, 1062640.5, 1063824.9, 1064926.9, 1065270.6, 1065392.3, 1066232.1, 1067493.2, 1068713.0, 1069604.4, 1070957.5, 1070624.1, 1070393.6, 1070380.4, 1070359.9, 1068133.3, 1065835.0, 1063721.5, 1060946.7, 1059518.6, 1055572.8, 1053347.0, 1051633.9, 1050395.3, 1048137.3, 1044862.1, 1041481.7, 1039339.6, 1036612.1, 1034686.5, 1033947.5, 1031267.0, 1029156.2, 1023041.4, 1022491.6, 1019763.5, 1018087.5, 1014178.3, 1011672.7, 1011367.8, 1007077.0, 1005434.2, 1003780.0, 1003315.9, 1001597.4, 999697.8, 999205.0, 998855.5, 998604.9, 998343.8, 992146.6, 989697.9, 987289.5, 986724.3, 986509.4