Major Bodies of Water in the City of St. Louis

A simple features data set containing polygons for the major bodies of water in the city.

data(stlHydro)

Format

A simple feature with 2 rows and 11 variables:

ANSICODE

ANSI identification number

HYDROID

Census identification number

FULLNAME

name of body of water

ALAND

area covered by land

INTPTLAT

interpolated latitude of polygon's centroid

INTPTLON

interpolated longitude of polygon's centroid

Shape_Leng

length of polygon

Shape_Area

area of polygon

geometry

simple features geometry

Examples

class(stlHydro)

#> [1] "sf"         "data.frame"

str(stlHydro)

#> Classes ‘sf’ and 'data.frame':	2 obs. of  11 variables:
#>  $ ANSICODE  : Factor w/ 0 levels: NA NA
#>  $ HYDROID   : Factor w/ 2 levels "1102212977721",..: 1 2
#>  $ FULLNAME  : Factor w/ 2 levels "Mississippi River",..: 1 2
#>  $ MTFCC     : Factor w/ 1 level "H3010": 1 1
#>  $ ALAND     : num  0 0
#>  $ AWATER    : num  10229167 440873
#>  $ INTPTLAT  : Factor w/ 2 levels "+38.5447301",..: 2 1
#>  $ INTPTLON  : Factor w/ 2 levels "-90.1753654",..: 1 2
#>  $ Shape_Leng: num  0.628 0.203
#>  $ Shape_Area: num  1.06e-03 4.56e-05
#>  $ geometry  :List of 2
#>   ..$ :List of 1
#>   .. ..$ : num [1:870, 1:2] 888812 888890 888957 889035 889073 ...
#>   .. ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ : num [1:319, 1:2] 871882 871890 871914 872015 872173 ...
#>   .. ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>   ..- attr(*, "n_empty")= int 0
#>   ..- attr(*, "crs")=List of 2
#>   .. ..$ epsg       : int NA
#>   .. ..$ proj4string: chr "+proj=tmerc +lat_0=35.83333333333334 +lon_0=-90.5 +k=0.9999333333333333 +x_0=250000 +y_0=0 +datum=NAD83 +units=us-ft +no_defs"
#>   .. ..- attr(*, "class")= chr "crs"
#>   ..- attr(*, "class")= chr  "sfc_POLYGON" "sfc"
#>   ..- attr(*, "precision")= num 0
#>   ..- attr(*, "bbox")=Class 'bbox'  Named num [1:4] 871882 982585 915267 1070614
#>   .. .. ..- attr(*, "names")= chr [1:4] "xmin" "ymin" "xmax" "ymax"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA
#>   ..- attr(*, "names")= chr  "ANSICODE" "HYDROID" "FULLNAME" "MTFCC" ...

head(stlHydro)

#>   ANSICODE       HYDROID          FULLNAME MTFCC ALAND   AWATER    INTPTLAT
#> 1     <NA> 1102212977721 Mississippi River H3010     0 10229167 +38.7658240
#> 2     <NA>  110451975812   River Des Peres H3010     0   440873 +38.5447301
#>      INTPTLON Shape_Leng   Shape_Area
#> 1 -90.1753654  0.6275771 1.058760e-03
#> 2 -90.2704658  0.2027437 4.556775e-05
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             geometry
#> 1 888811.7, 888889.5, 888957.0, 889035.1, 889072.9, 889144.6, 889194.2, 889297.9, 889351.5, 889445.3, 889517.1, 889682.6, 889921.6, 889939.3, 890129.1, 890232.7, 890360.0, 890423.8, 890501.6, 890832.2, 890899.8, 890976.7, 890983.3, 891071.0, 891089.1, 891072.7, 891082.3, 891157.9, 891133.4, 891105.2, 891123.3, 891159.0, 891208.6, 891290.6, 891392.5, 891430.4, 891474.3, 891526.3, 891547.5, 891615.5, 891649.2, 891665.1, 891703.8, 891741.5, 891835.0, 891922.8, 891930.7, 891956.0, 891993.8, 892045.8, 892057.6, 892099.5, 892121.4, 892324.7, 892420.1, 892507.5, 892683.3, 892728.9, 892760.4, 892796.2, 892883.7, 892933.7, 892991.5, 893088.8, 893100.9, 893130.0, 893175.5, 893221.0, 893332.1, 893377.9, 893491.5, 893629.0, 893637.2, 893658.5, 893688.7, 893804.2, 893891.9, 893951.2, 894045.0, 894075.4, 894637.4, 894721.0, 894806.5, 894937.8, 894991.5, 895227.8, 895351.5, 895416.9, 895571.9, 895667.3, 895747.0, 895870.4, 895958.0, 896041.2, 896175.2, 896221.8, 896341.5, 896596.4, 896817.8, 896997.2, 897363.7, 897470.6, 897546.4, 897653.6, 897745.1, 897867.7, 898408.1, 898809.5, 898994.8, 899760.2, 899838.1, 899941.1, 900027.8, 900137.4, 900187.2, 900436.8, 900592.3, 900601.5, 900739.0, 900982.3, 901317.6, 901389.8, 901401.0, 901487.1, 901622.6, 901829.9, 902021.1, 902318.2, 902647.3, 903175.7, 903508.6, 903687.0, 903773.4, 904129.6, 904286.4, 904443.1, 904583.9, 904644.6, 904899.7, 905000.8, 905047.3, 905069.2, 905127.8, 905178.8, 905533.3, 905794.0, 905913.3, 906018.8, 906138.1, 906156.3, 906303.4, 906484.7, 906615.0, 906755.3, 906898.6, 906974.3, 907067.5, 907115.2, 907157.0, 907266.3, 907433.5, 907852.4, 907945.6, 907991.5, 908148.1, 908564.4, 908955.7, 909039.0, 909154.5, 909277.5, 909357.4, 909370.4, 909477.7, 909583.4, 909586.5, 909636.7, 909649.4, 909673.3, 909691.7, 909707.8, 909718.6, 909729.3, 909770.4, 909786.2, 909918.5, 910007.1, 910051.2, 910065.6, 910086.3, 910094.4, 910102.4, 910108.9, 910138.8, 910328.9, 910419.0, 910839.3, 910856.5, 910872.2, 910885.9, 911003.1, 911026.0, 911130.6, 911189.6, 911192.4, 911199.7, 911202.2, 911229.0, 911268.3, 911301.1, 911301.6, 911310.3, 911316.6, 911317.7, 911318.0, 911318.8, 911321.2, 911322.0, 911332.1, 911334.3, 911341.5, 911355.1, 911398.1, 911398.9, 911433.4, 911437.6, 911456.9, 911482.5, 911532.2, 911552.0, 911559.6, 911554.5, 911557.5, 911561.1, 911548.7, 911532.0, 911513.5, 911485.0, 911472.7, 911466.2, 911465.7, 911461.1, 911460.7, 911457.0, 911452.4, 911436.8, 911425.0, 911412.9, 911368.8, 911327.7, 911323.1, 911333.1, 911332.8, 911304.8, 911295.9, 911249.0, 911196.0, 911190.9, 911188.2, 911183.0, 911177.9, 911132.7, 911148.8, 911158.6, 911152.0, 911147.7, 911135.3, 911108.5, 911112.5, 911164.5, 911154.2, 911113.8, 911065.2, 911052.9, 911044.0, 911011.3, 910930.4, 910899.9, 910902.6, 910901.3, 910853.2, 910812.2, 910802.5, 910704.5, 910690.0, 910640.8, 910550.9, 910505.5, 910466.9, 910435.8, 910423.4, 910396.2, 910330.6, 910295.2, 910240.2, 910226.0, 910177.1, 910191.1, 910202.3, 910193.4, 910178.0, 910140.3, 910109.2, 910100.8, 910105.1, 910096.6, 910092.1, 910070.8, 910040.2, 910017.8, 909973.4, 909947.2, 909941.0, 909938.6, 909916.4, 909898.3, 909869.8, 909862.0, 909815.8, 909767.7, 909765.6, 909761.4, 909755.9, 909747.3, 909701.0, 909636.5, 909561.7, 909509.2, 909494.6, 909482.6, 909454.2, 909406.3, 909387.9, 909361.6, 909343.0, 909328.7, 909304.6, 909288.2, 909259.8, 909223.3, 909188.6, 909160.1, 909148.7, 909113.3, 909112.9, 909078.2, 909041.7, 908973.5, 908967.2, 908898.2, 908868.0, 908825.7, 908799.5, 908747.1, 908703.1, 908657.0, 908626.8, 908566.6, 908528.1, 908473.9, 908431.9, 908389.7, 908311.4, 908251.2, 908142.7, 908112.8, 908076.6, 908034.2, 907992.3, 907956.2, 907861.3, 907744.4, 907710.3, 907626.0, 907575.8, 907517.2, 907412.8, 907329.9, 907211.5, 906899.9, 906735.1, 906676.9, 906634.5, 906580.3, 906536.1, 906457.9, 906348.4, 906293.0, 906255.0, 906209.0, 906191.0, 906192.9, 906164.9, 906146.8, 906144.6, 906106.6, 906050.7, 905964.2, 905945.9, 905892.0, 905823.9, 905595.5, 905464.9, 905378.6, 905280.4, 905161.9, 905007.6, 904871.2, 904736.6, 904647.2, 904465.9, 904258.9, 904186.6, 904071.8, 903983.4, 903905.3, 903789.3, 903715.0, 903716.8, 903790.2, 903822.4, 903862.4, 903872.3, 903803.5, 903685.1, 903596.8, 903408.0, 903249.5, 903205.0, 903140.6, 903001.9, 902961.6, 902933.6, 902869.5, 902843.4, 902827.0, 902736.5, 902710.0, 902665.8, 902635.5, 902591.2, 902516.9, 902466.4, 902422.9, 902372.2, 902346.1, 902383.9, 902473.6, 902475.4, 902459.3, 902409.2, 902317.2, 902096.8, 902040.7, 901994.1, 901950.4, 901912.3, 901856.1, 901801.9, 901741.4, 901716.8, 901648.3, 901636.1, 901549.2, 901522.9, 901512.6, 901512.5, 901512.2, 901494.1, 901457.9, 901433.7, 901425.5, 901404.9, 901344.5, 901318.2, 901316.0, 901289.4, 901264.7, 901230.4, 901184.1, 901123.5, 901076.8, 901022.2, 900985.7, 900976.8, 900954.1, 900959.2, 900973.0, 900988.6, 900992.2, 901001.8, 901021.4, 901057.0, 901062.8, 901082.3, 901139.0, 901172.4, 901196.0, 901205.6, 901209.3, 901205.0, 901194.3, 901194.0, 901194.0, 901203.5, 901217.3, 901250.8, 901272.5, 901288.2, 901310.8, 901365.2, 901410.7, 901444.3, 901495.7, 901541.4, 901586.9, 901603.6, 901606.7, 901610.5, 901654.0, 901657.8, 901635.4, 901621.3, 901591.2, 901551.2, 901499.2, 901433.3, 901347.0, 901297.3, 901227.1, 901108.9, 901070.9, 901032.9, 900956.9, 900866.7, 900830.7, 900874.6, 901108.5, 901156.5, 901200.5, 901362.7, 901467.0, 901545.0, 901579.0, 901630.9, 901669.2, 901687.2, 901703.1, 901715.0, 901732.8, 901758.6, 901766.5, 901757.8, 901755.2, 901738.7, 901738.3, 901742.1, 901809.4, 901829.2, 901892.6, 901952.4, 902029.8, 902105.4, 902149.1, 902264.7, 902358.3, 902461.7, 902573.5, 902641.1, 902698.8, 902724.6, 902782.3, 902860.2, 902951.6, 903037.2, 903113.1, 903178.7, 903250.2, 903311.8, 903401.6, 903495.1, 903612.6, 903698.5, 903801.9, 903913.8, 903949.5, 904035.0, 904128.9, 904186.6, 904236.4, 904332.0, 904451.7, 904573.2, 904730.8, 904826.7, 904936.3, 905103.9, 905271.4, 905398.8, 905494.7, 905570.3, 905697.7, 905783.9, 905901.2, 906046.8, 906194.4, 906325.8, 906443.5, 906589.1, 906666.8, 906766.4, 906866.2, 906961.8, 907115.5, 907213.4, 907272.0, 907352.8, 907462.7, 907562.3, 907686.0, 907813.9, 907969.2, 908100.9, 908230.7, 908479.9, 908529.7, 908607.0, 908701.2, 908733.1, 908777.0, 908827.0, 908870.9, 908950.6, 909030.5, 909076.2, 909130.0, 909287.7, 909405.5, 909618.8, 909694.5, 909782.4, 909903.2, 909908.8, 909652.7, 909991.4, 910230.1, 910436.7, 910547.2, 910759.2, 910825.3, 910819.0, 910846.9, 910858.6, 910890.4, 910950.0, 910997.3, 911013.3, 911039.0, 911070.5, 911108.0, 911137.8, 911181.6, 911211.0, 911230.1, 911278.9, 911429.6, 911495.9, 911518.4, 911633.4, 911678.8, 911717.6, 911790.1, 911812.8, 911816.5, 911827.2, 911830.6, 911835.7, 911885.2, 911932.8, 911984.6, 912001.5, 912076.9, 912088.5, 912120.1, 912387.2, 912431.1, 912468.7, 912528.0, 912529.7, 912583.8, 912649.3, 912752.7, 912776.2, 912784.7, 914922.8, 915266.8, 914782.6, 914549.3, 914174.9, 913998.2, 913716.4, 913693.7, 913618.4, 913319.3, 913000.9, 912922.8, 912887.1, 912787.3, 912629.6, 911720.3, 910442.6, 909318.4, 907510.8, 906415.7, 905713.7, 904810.8, 904299.1, 903934.0, 903312.6, 903006.3, 902965.4, 902897.3, 902892.2, 902631.6, 902608.2, 902542.9, 902404.1, 902398.4, 902234.9, 902105.7, 902420.3, 902704.0, 902896.2, 903495.6, 904108.4, 905107.2, 905365.6, 906190.5, 906514.6, 906740.6, 906918.1, 907067.0, 907524.5, 908180.3, 908478.0, 908797.5, 909095.8, 909331.4, 909527.0, 909625.4, 909653.4, 909671.9, 909968.3, 910166.6, 910287.3, 910427.7, 910525.7, 910669.2, 910797.6, 911120.5, 911168.5, 911212.0, 911411.9, 911561.6, 911789.1, 911801.9, 911784.0, 911858.9, 912006.5, 912022.2, 912029.8, 912108.3, 912213.1, 912236.5, 912267.2, 912278.5, 912319.3, 912287.9, 912216.2, 912063.8, 912062.7, 912060.5, 912039.8, 911970.2, 911961.0, 911939.0, 911802.5, 911377.4, 910935.1, 910904.3, 910895.5, 910893.0, 910888.5, 910884.8, 910882.0, 910850.4, 910814.0, 910609.0, 910582.8, 910560.9, 910554.4, 910392.9, 910260.0, 910232.3, 910207.8, 910179.0, 910142.8, 909939.9, 909672.6, 909459.4, 909208.4, 908727.1, 908104.6, 907817.6, 907539.0, 907147.6, 906697.6, 906431.0, 906059.2, 905665.8, 905496.1, 905046.5, 904643.2, 904389.8, 902848.2, 902621.0, 901865.1, 900817.8, 900520.6, 900144.4, 899937.5, 899653.8, 899379.2, 899096.8, 897597.4, 897483.9, 896741.6, 896711.2, 896218.0, 896073.6, 896046.9, 896003.7, 895835.7, 895640.1, 895624.5, 895495.4, 895432.6, 895402.5, 895243.3, 895180.1, 895055.2, 894679.4, 894610.2, 894553.8, 894394.4, 894195.9, 893916.0, 893756.5, 893230.8, 892509.1, 892204.5, 891947.3, 891591.2, 891045.2, 890361.9, 889828.5, 889615.2, 889600.6, 889215.1, 888758.4, 888811.7, 983116.6, 983234.8, 983370.8, 983445.3, 983539.4, 983611.7, 983779.7, 983915.8, 984095.9, 984231.9, 984429.9, 984680.1, 985063.1, 985086.1, 985335.4, 985533.4, 985853.5, 985957.8, 986059.6, 986708.4, 986882.3, 986995.0, 987004.5, 987126.3, 987220.4, 987360.5, 987514.2, 987772.6, 987978.3, 988042.3, 988108.3, 988226.4, 988274.2, 988312.7, 988390.5, 988444.5, 988506.9, 988724.5, 988814.9, 988852.9, 988961.2, 989004.9, 989086.6, 989190.5, 989297.1, 989413.1, 989461.9, 989623.3, 989685.4, 989701.9, 989705.6, 989745.7, 989787.7, 989984.2, 990158.5, 990274.9, 990333.3, 990465.6, 990569.5, 990631.5, 990739.9, 990774.3, 990850.2, 991050.4, 991105.0, 991238.4, 991420.6, 991510.7, 991767.0, 991849.1, 991923.7, 992070.8, 992080.0, 992128.5, 992198.1, 992306.6, 992424.8, 992576.9, 992701.3, 992748.4, 993616.7, 993796.8, 993893.2, 994079.7, 994203.7, 994632.2, 994790.6, 994920.8, 995211.2, 995321.8, 995451.7, 995602.1, 995748.3, 995942.7, 996168.1, 996265.9, 996494.2, 996835.8, 997110.7, 997255.4, 997632.7, 997862.4, 997916.2, 997919.8, 997933.9, 998034.0, 998693.0, 999149.7, 999317.1, 999949.4, 1000013.8, 1000072.3, 1000122.5, 1000200.8, 1000257.0, 1000529.8, 1000700.3, 1000710.2, 1000874.5, 1001045.3, 1001236.5, 1001275.7, 1001281.9, 1001328.8, 1001431.1, 1001674.0, 1001884.3, 1002125.3, 1002360.4, 1002778.0, 1003072.9, 1003231.9, 1003309.1, 1003628.1, 1003768.5, 1003908.4, 1004033.8, 1004088.3, 1004313.1, 1004418.6, 1004467.6, 1004494.2, 1004565.8, 1004627.9, 1005048.9, 1005373.9, 1005518.1, 1005696.2, 1005824.8, 1005844.9, 1006037.3, 1006289.5, 1006464.4, 1006652.4, 1006876.8, 1006975.1, 1007119.2, 1007207.5, 1007291.4, 1007493.9, 1007766.1, 1008617.5, 1008797.7, 1008819.7, 1009260.2, 1010212.8, 1011110.0, 1011308.1, 1011460.7, 1011763.0, 1011958.5, 1011990.9, 1012221.5, 1012552.9, 1012563.4, 1012762.8, 1012812.7, 1012907.9, 1012961.1, 1013009.6, 1013041.3, 1013073.0, 1013194.1, 1013251.7, 1013736.1, 1013985.2, 1014110.6, 1014158.0, 1014226.9, 1014253.5, 1014279.7, 1014301.6, 1014399.7, 1015027.8, 1015327.1, 1016572.3, 1016632.8, 1016714.8, 1016804.8, 1017271.3, 1017363.2, 1017763.0, 1018037.1, 1018048.8, 1018083.8, 1018095.8, 1018219.3, 1018403.4, 1018592.5, 1018597.6, 1018720.4, 1018807.4, 1018824.2, 1018828.6, 1018836.6, 1018873.7, 1018886.5, 1019031.5, 1019070.4, 1019217.2, 1019401.2, 1019613.3, 1019616.6, 1019796.2, 1019817.4, 1020029.4, 1020135.5, 1020227.8, 1020281.7, 1020383.7, 1020721.7, 1021009.4, 1021121.2, 1021237.3, 1021441.2, 1021581.0, 1021730.9, 1021814.7, 1021938.8, 1022096.5, 1022278.6, 1022292.4, 1022379.5, 1022461.7, 1022522.1, 1022638.6, 1022753.3, 1022929.1, 1023083.0, 1023110.6, 1023267.3, 1023366.0, 1023435.8, 1023455.4, 1023558.3, 1023750.4, 1023833.1, 1023876.4, 1023956.1, 1024036.6, 1024461.4, 1024535.4, 1024581.0, 1024677.8, 1024748.8, 1024960.0, 1025183.5, 1025286.6, 1025336.7, 1025432.4, 1025715.3, 1026054.2, 1026139.3, 1026173.9, 1026197.5, 1026230.7, 1026283.4, 1026599.1, 1026709.1, 1026987.9, 1027224.9, 1027279.1, 1027310.8, 1027367.6, 1027481.4, 1027638.0, 1027783.2, 1027969.9, 1028120.9, 1028256.7, 1028365.1, 1028669.4, 1028832.8, 1029042.7, 1029098.4, 1029289.8, 1029444.6, 1029526.2, 1029619.0, 1029686.4, 1029779.1, 1029861.7, 1029885.7, 1029978.6, 1030042.3, 1030098.3, 1030236.6, 1030416.4, 1030514.3, 1030640.2, 1030710.0, 1030775.9, 1030877.9, 1030955.7, 1030994.6, 1031056.8, 1031073.5, 1031137.5, 1031187.2, 1031217.8, 1031273.1, 1031288.4, 1031313.2, 1031409.1, 1031572.8, 1031814.4, 1031970.1, 1032064.0, 1032153.9, 1032192.0, 1032249.8, 1032301.8, 1032399.7, 1032573.3, 1032667.2, 1032705.0, 1032818.9, 1032947.0, 1033088.6, 1033288.4, 1033464.2, 1033523.2, 1033706.2, 1033824.2, 1034040.1, 1034191.8, 1034426.5, 1034449.0, 1034748.5, 1034804.5, 1034920.2, 1034970.3, 1035114.0, 1035195.8, 1035257.6, 1035315.4, 1035379.3, 1035457.1, 1035542.8, 1035634.5, 1035700.6, 1035796.1, 1035873.5, 1036043.2, 1036127.6, 1036177.0, 1036242.8, 1036298.3, 1036346.3, 1036462.2, 1036667.9, 1036709.7, 1036809.2, 1036889.5, 1036999.3, 1037153.0, 1037265.2, 1037426.2, 1037909.1, 1038174.4, 1038278.0, 1038395.9, 1038479.8, 1038543.4, 1038625.1, 1038780.6, 1038858.7, 1038868.5, 1038884.3, 1038908.3, 1038930.2, 1038940.3, 1038978.1, 1039030.2, 1039039.9, 1039013.8, 1039105.3, 1039127.1, 1039191.0, 1039257.1, 1039435.9, 1039565.5, 1039679.2, 1039795.0, 1039886.4, 1040025.8, 1040187.4, 1040326.8, 1040432.1, 1040645.7, 1040934.2, 1041034.8, 1041248.2, 1041379.8, 1041429.8, 1041493.1, 1041522.8, 1041574.9, 1041640.3, 1041669.2, 1041743.2, 1041779.3, 1041963.0, 1042118.5, 1042238.4, 1042477.4, 1042667.0, 1042774.7, 1042916.5, 1043168.1, 1043205.8, 1043211.6, 1043255.4, 1043315.4, 1043419.2, 1043634.9, 1043706.9, 1043782.5, 1043890.6, 1043988.4, 1044116.0, 1044197.8, 1044241.7, 1044293.2, 1044349.2, 1044395.3, 1044461.8, 1044517.5, 1044559.4, 1044599.3, 1044652.9, 1044762.2, 1044814.1, 1044834.7, 1044853.9, 1044891.6, 1044991.2, 1045081.0, 1045250.9, 1045368.8, 1045564.5, 1045616.2, 1045866.1, 1045969.8, 1046067.8, 1046080.9, 1046193.8, 1046239.6, 1046301.4, 1046381.5, 1046455.4, 1046555.1, 1046715.1, 1046813.0, 1046887.0, 1047098.8, 1047214.6, 1047336.5, 1047468.1, 1047683.9, 1047817.8, 1048029.6, 1048221.4, 1048421.3, 1048664.9, 1048962.8, 1049558.6, 1049694.5, 1049822.7, 1049938.6, 1050062.8, 1050180.6, 1050223.9, 1050374.8, 1050762.8, 1050942.8, 1051088.9, 1051188.7, 1051300.9, 1051400.7, 1051522.7, 1051632.6, 1051642.5, 1051770.7, 1051842.5, 1051974.8, 1052079.0, 1052156.6, 1052238.6, 1052436.9, 1052571.1, 1052697.2, 1052873.3, 1053045.3, 1053169.3, 1053227.6, 1053239.3, 1053279.3, 1053421.5, 1053487.4, 1053523.4, 1053581.3, 1053615.4, 1053639.0, 1053653.0, 1053680.8, 1053708.3, 1053727.0, 1053754.1, 1053763.6, 1053779.5, 1053803.4, 1053830.9, 1053856.8, 1053892.8, 1053904.9, 1053885.6, 1053877.7, 1053861.8, 1053794.2, 1053752.7, 1053709.2, 1053699.1, 1053687.3, 1053661.5, 1053675.4, 1053695.5, 1053725.8, 1053777.6, 1053821.7, 1053859.6, 1054065.7, 1054185.5, 1054345.3, 1054471.3, 1054533.3, 1054735.6, 1054795.8, 1054953.6, 1055083.8, 1055214.1, 1055302.5, 1055382.4, 1055531.0, 1055640.9, 1055755.2, 1055857.5, 1055941.9, 1056003.9, 1056054.3, 1056148.4, 1056228.4, 1056350.7, 1056451.1, 1056527.1, 1056619.5, 1056729.7, 1056839.9, 1056938.1, 1057036.4, 1057126.7, 1057214.8, 1057335.3, 1057425.6, 1057475.6, 1057588.0, 1057666.3, 1057730.2, 1057778.4, 1057854.9, 1057947.0, 1058051.6, 1058175.9, 1058246.1, 1058322.6, 1058455.0, 1058605.6, 1058710.1, 1058792.4, 1058848.4, 1058970.8, 1059031.9, 1059115.7, 1059232.0, 1059370.5, 1059484.9, 1059593.1, 1059721.8, 1059767.9, 1059832.3, 1059930.6, 1060004.9, 1060117.2, 1060181.6, 1060215.3, 1060261.8, 1060334.3, 1060408.6, 1060493.1, 1060561.3, 1060683.8, 1060776.4, 1060856.6, 1061051.6, 1061093.6, 1061147.8, 1061214.0, 1061230.1, 1061238.3, 1061290.5, 1061310.7, 1061352.9, 1061409.2, 1061467.3, 1061503.5, 1061602.0, 1061680.4, 1061861.0, 1061913.0, 1061963.5, 1062087.1, 1062110.7, 1062109.9, 1062470.9, 1062793.6, 1063180.0, 1063636.4, 1064188.8, 1064678.9, 1064768.8, 1064884.7, 1064948.9, 1065005.1, 1065141.9, 1065251.3, 1065333.3, 1065397.5, 1065531.2, 1065627.5, 1065669.5, 1065787.6, 1065951.6, 1066201.5, 1066505.8, 1066967.4, 1067170.1, 1067240.1, 1067544.6, 1067686.8, 1067763.8, 1067909.0, 1067946.6, 1067952.4, 1067969.9, 1067976.1, 1067984.9, 1068107.1, 1068285.3, 1068532.4, 1068613.3, 1068821.6, 1068903.5, 1069019.5, 1069834.4, 1069918.3, 1070018.6, 1070158.6, 1070161.9, 1070272.8, 1070367.0, 1070551.3, 1070596.9, 1070613.7, 1070330.8, 1070286.5, 1069448.2, 1068928.0, 1068215.0, 1067881.9, 1067351.3, 1067321.0, 1067183.4, 1066624.1, 1066027.5, 1065865.5, 1065791.5, 1065383.6, 1063933.2, 1062196.4, 1060910.0, 1060196.0, 1059095.9, 1058249.9, 1057476.2, 1056569.0, 1055572.8, 1055031.9, 1054252.3, 1053414.8, 1053347.0, 1053069.6, 1053047.7, 1051992.6, 1051897.8, 1051633.9, 1050395.3, 1049765.9, 1049143.4, 1048223.1, 1047438.9, 1046419.7, 1045462.5, 1044170.8, 1042918.1, 1041427.5, 1041054.7, 1040099.2, 1039777.6, 1039555.1, 1039380.9, 1039249.2, 1038753.2, 1037787.0, 1037290.2, 1036679.4, 1036066.8, 1035719.5, 1035288.2, 1034885.4, 1034736.2, 1034647.0, 1033809.3, 1033179.6, 1032612.2, 1032128.7, 1031876.7, 1031584.0, 1031176.2, 1029454.0, 1029198.9, 1029028.6, 1028364.3, 1027765.4, 1026568.7, 1026019.6, 1025673.5, 1025135.5, 1024267.8, 1024174.3, 1024128.1, 1023652.7, 1023004.5, 1022770.8, 1022485.4, 1022370.3, 1021894.8, 1020751.9, 1019773.1, 1018751.4, 1018744.8, 1018730.6, 1018595.8, 1018148.7, 1018090.1, 1018002.6, 1017179.1, 1015609.0, 1014146.4, 1014044.7, 1014015.9, 1014006.1, 1013989.7, 1013976.5, 1013967.1, 1013855.5, 1013726.1, 1013003.6, 1012911.4, 1012792.6, 1012757.6, 1012375.0, 1011817.0, 1011705.4, 1011608.1, 1011508.6, 1011404.0, 1010918.5, 1010340.8, 1009859.0, 1009295.8, 1008290.1, 1006955.5, 1006477.5, 1006013.7, 1005462.1, 1004906.0, 1004609.0, 1004243.6, 1003772.9, 1003566.6, 1003124.1, 1002840.2, 1002614.3, 1001373.4, 1001086.1, 1000381.6, 999528.4, 999311.2, 999106.8, 998982.7, 998829.6, 998592.8, 997929.9, 996257.1, 996130.1, 995303.0, 995240.7, 994231.2, 993935.1, 993880.4, 993791.0, 993447.5, 993046.7, 993015.3, 992750.9, 992621.5, 992560.2, 992234.2, 992105.4, 991908.4, 991317.0, 991208.0, 991120.4, 990879.6, 990368.8, 990157.2, 989916.4, 989453.5, 988676.5, 987943.4, 987149.5, 986604.8, 985670.3, 984362.6, 983275.5, 982612.9, 982585.2, 982796.5, 983047.3, 983116.6
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              871882.3, 871890.2, 871913.8, 872015.3, 872172.9, 872446.7, 872500.1, 872723.7, 872736.7, 872750.1, 872763.1, 872768.5, 872770.2, 872777.3, 872782.1, 872785.5, 872787.2, 872792.3, 872794.0, 872795.4, 872800.8, 872805.0, 872808.3, 872809.7, 872811.7, 872816.2, 872860.0, 872862.9, 872864.1, 872867.6, 872868.5, 872871.4, 872874.3, 872874.9, 872877.6, 872825.8, 872792.7, 872752.8, 872734.0, 872495.2, 872423.1, 872330.9, 872225.2, 872163.6, 872116.5, 872081.5, 872011.5, 871980.5, 871962.3, 871971.5, 872026.7, 872029.5, 872056.9, 872079.6, 872146.9, 872157.6, 872169.6, 872192.1, 872258.5, 872280.7, 872303.4, 872371.5, 872379.0, 872404.8, 872649.4, 872655.7, 873094.8, 873435.3, 873605.8, 873698.4, 873866.7, 873871.3, 874097.3, 874263.8, 874299.8, 874322.2, 874380.7, 874380.2, 874371.4, 874408.9, 874442.7, 874621.6, 874832.5, 875018.0, 875240.2, 875277.5, 875729.8, 876054.4, 876411.3, 876573.4, 876948.2, 877018.2, 877148.6, 877236.5, 877320.8, 877912.0, 878436.6, 878677.1, 878815.0, 879111.9, 879370.3, 879584.5, 879782.5, 879895.5, 879996.5, 880276.5, 880524.5, 881036.4, 881522.4, 881668.4, 881798.6, 881998.7, 882255.0, 882283.1, 882369.7, 882437.3, 882800.7, 882936.5, 883046.3, 883063.5, 883350.6, 883489.3, 883506.8, 883680.8, 883827.3, 883918.8, 884122.7, 884492.7, 884668.7, 884824.6, 884894.4, 885068.7, 885229.9, 885313.0, 885412.7, 885559.2, 885747.8, 885808.0, 885840.3, 885838.2, 885837.2, 885836.1, 885833.0, 885763.7, 885744.2, 885742.9, 885683.1, 885687.7, 885684.7, 885658.4, 885712.7, 885732.6, 885810.7, 885838.6, 885843.8, 885849.0, 885950.5, 886103.5, 886142.5, 886184.1, 886198.6, 886197.8, 886179.0, 886167.4, 886153.9, 886200.6, 886273.6, 886282.8, 886408.9, 886537.3, 886541.0, 886683.5, 886787.5, 886949.8, 887035.8, 887106.1, 887172.1, 887250.3, 887342.3, 887414.9, 887570.6, 887752.8, 887910.9, 887923.0, 887951.2, 888043.2, 888060.7, 888151.3, 888240.2, 888312.1, 888559.0, 888758.4, 888599.0, 888253.2, 888204.8, 888167.6, 888164.2, 888124.3, 888111.7, 888049.7, 888017.0, 887972.6, 887947.4, 887851.6, 887639.4, 887354.7, 886686.6, 886682.0, 886369.5, 886327.7, 886205.3, 886201.9, 886162.6, 886026.9, 885968.9, 885912.2, 885922.2, 885957.4, 885972.9, 885976.5, 885837.1, 885705.4, 885693.3, 885496.6, 885474.1, 885473.8, 885451.8, 885386.0, 885367.1, 885363.6, 885368.2, 885381.0, 885475.1, 885484.4, 885483.4, 885481.8, 885495.4, 885560.9, 885586.5, 885637.4, 885672.6, 885676.2, 885676.5, 885680.4, 885684.9, 885689.1, 885695.8, 885659.5, 885620.8, 885574.3, 885513.4, 885301.2, 885067.0, 884677.0, 884227.1, 883903.1, 883677.2, 883463.1, 883378.0, 883214.6, 882553.8, 882194.2, 882093.0, 881642.6, 881340.8, 881196.5, 881022.8, 880776.8, 880606.9, 880258.8, 879908.7, 879478.6, 879250.4, 879054.2, 878805.6, 878721.4, 878066.5, 877953.6, 877303.1, 877097.3, 876672.0, 876205.3, 875856.4, 875707.3, 875495.4, 875294.9, 875111.4, 875029.9, 874980.2, 874680.7, 874603.7, 874491.6, 874471.5, 874308.7, 874302.0, 874239.8, 874241.1, 874244.7, 874196.4, 873999.4, 873915.1, 873730.7, 873600.6, 873333.7, 872944.7, 872714.2, 872530.8, 872381.5, 872275.9, 872267.0, 872246.9, 872163.9, 872136.1, 872110.2, 872001.3, 871986.6, 871914.9, 871886.6, 871882.3, 1005975.7, 1006032.2, 1006205.6, 1006371.8, 1006592.1, 1006845.0, 1006894.7, 1007128.9, 1007171.9, 1007216.0, 1007260.5, 1007277.6, 1007283.8, 1007307.1, 1007331.1, 1007349.4, 1007357.7, 1007382.9, 1007391.3, 1007398.9, 1007425.1, 1007447.0, 1007506.0, 1007529.7, 1007565.0, 1007596.3, 1007422.0, 1007387.0, 1007372.1, 1007327.3, 1007312.7, 1007276.0, 1007239.2, 1007233.4, 1007197.3, 1007079.2, 1007027.4, 1006966.2, 1006937.0, 1006730.8, 1006668.0, 1006588.4, 1006451.3, 1006372.1, 1006310.9, 1006266.0, 1006094.3, 1006018.5, 1005973.7, 1005929.3, 1005659.9, 1005650.4, 1005560.9, 1005487.0, 1005265.4, 1005230.4, 1005191.8, 1005121.2, 1004909.4, 1004839.2, 1004767.1, 1004550.9, 1004527.3, 1004487.3, 1004105.7, 1004095.9, 1003466.8, 1003020.3, 1002796.3, 1002660.3, 1002413.0, 1002405.3, 1002021.2, 1001673.8, 1001559.5, 1001488.2, 1001262.1, 1001240.3, 1000926.3, 1000610.7, 1000322.7, 999789.2, 999287.8, 998892.7, 998419.0, 998353.1, 997556.2, 997107.1, 996657.8, 996492.1, 996097.0, 996015.2, 995851.6, 995801.9, 995685.9, 995053.2, 994566.7, 994305.0, 994156.4, 993836.2, 993612.8, 993479.3, 993409.5, 993385.7, 993364.1, 993324.7, 993293.2, 993254.3, 993169.1, 993119.6, 993079.8, 992972.1, 992784.8, 992759.4, 992682.0, 992621.4, 992281.7, 992155.0, 992052.5, 992037.6, 991786.3, 991665.0, 991649.8, 991556.2, 991535.4, 991522.6, 991516.9, 991527.6, 991503.6, 991482.5, 991459.8, 991403.0, 991308.4, 991259.4, 991168.2, 991034.2, 990746.6, 990546.8, 990438.8, 990356.8, 990319.3, 990282.2, 990164.9, 989724.8, 989528.8, 989261.2, 989147.0, 988775.2, 988725.3, 988290.4, 987948.6, 987886.4, 987655.3, 987589.5, 987570.2, 987551.6, 987300.2, 986866.5, 986706.0, 986535.0, 986335.1, 986321.6, 986077.2, 985897.3, 985603.3, 985307.4, 985181.2, 985165.6, 985058.1, 984982.3, 984980.1, 984852.3, 984759.0, 984625.0, 984517.4, 984438.2, 984363.7, 984306.7, 984240.0, 984165.1, 984004.6, 983884.8, 983819.3, 983789.9, 983719.7, 983627.8, 983615.4, 983561.4, 983555.1, 983518.1, 983276.2, 983047.3, 983099.7, 983280.5, 983312.1, 983310.5, 983313.4, 983353.4, 983366.1, 983461.7, 983483.1, 983512.9, 983529.6, 983531.5, 983716.3, 983964.3, 984522.7, 984526.7, 984793.2, 984828.8, 984933.4, 984937.0, 984977.7, 985119.0, 985224.9, 985490.9, 985574.7, 985908.7, 986122.6, 986426.6, 987010.1, 987418.0, 987452.9, 988001.2, 988192.7, 988328.6, 988420.6, 988442.7, 988458.3, 988468.5, 988478.7, 988488.2, 988494.6, 988660.4, 988706.2, 988810.7, 988962.6, 989200.6, 989344.5, 989930.6, 990261.4, 990294.1, 990299.2, 990337.5, 990381.6, 990426.0, 990486.8, 990698.3, 990755.4, 990823.0, 990912.1, 991151.6, 991286.8, 991378.0, 991389.3, 991406.7, 991426.2, 991511.6, 991581.3, 991715.0, 992319.7, 992640.5, 992730.5, 992997.5, 993078.7, 993126.5, 993086.4, 993121.9, 993141.5, 993180.8, 993245.9, 993393.1, 993518.6, 993676.2, 993987.4, 994075.7, 994761.8, 994878.5, 995550.1, 995773.6, 996232.6, 996725.4, 997152.6, 997367.9, 997673.7, 997999.2, 998376.8, 998544.5, 998650.8, 999289.3, 999486.9, 999775.1, 999824.9, 1000302.4, 1000345.3, 1000736.0, 1001064.1, 1001248.0, 1001497.7, 1001959.1, 1002090.0, 1002376.6, 1002564.3, 1002949.0, 1003510.5, 1003853.9, 1004101.9, 1004303.3, 1004498.3, 1004515.0, 1004562.7, 1004759.6, 1004825.4, 1004886.6, 1005330.3, 1005386.7, 1005661.9, 1005797.6, 1005975.7