City of St. Louis Boundary

A simple features data set containing a single polygon with the extent of the City of St. Louis's boundary.

data(stl_sf_boundary)

Format

A simple feature with 1 rows and 5 variables:

GEOID

county FIPS code

NAME

polygon name

LAND_SQMI

land area of city, square miles

WATER_SQMI

water area of city, square miles

geometry

simple features geometry

Examples

class(stl_sf_boundary)

#> [1] "sf"         "data.frame"

str(stl_sf_boundary)

#> Classes ‘sf’ and 'data.frame':	1 obs. of  5 variables:
#>  $ GEOID     : chr "29510"
#>  $ NAME      : chr "St. Louis City"
#>  $ LAND_SQMI : num 62
#>  $ WATER_SQMI: num 4.12
#>  $ geometry  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:660, 1:2] -90.3 -90.3 -90.3 -90.3 -90.3 ...
#>   .. ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>   ..- attr(*, "n_empty")= int 0
#>   ..- attr(*, "crs")=List of 2
#>   .. ..$ epsg       : int 4269
#>   .. ..$ proj4string: chr "+proj=longlat +datum=NAD83 +no_defs"
#>   .. ..- attr(*, "class")= chr "crs"
#>   ..- attr(*, "class")= chr  "sfc_POLYGON" "sfc"
#>   ..- attr(*, "precision")= num 0
#>   ..- attr(*, "bbox")= 'bbox' Named num  -90.3 38.5 -90.2 38.8
#>   .. ..- attr(*, "names")= chr  "xmin" "ymin" "xmax" "ymax"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA
#>   ..- attr(*, "names")= chr  "GEOID" "NAME" "LAND_SQMI" "WATER_SQMI"

head(stl_sf_boundary)

#>   GEOID           NAME LAND_SQMI WATER_SQMI
#> 1 29510 St. Louis City   61.9532   4.119725
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             geometry
#> 1 -90.32052, -90.32050, -90.31964, -90.31959, -90.31941, -90.31904, -90.31897, -90.31893, -90.31891, -90.31882, -90.31873, -90.31865, -90.31858, -90.31843, -90.31821, -90.31793, -90.31746, -90.31711, -90.31706, -90.31664, -90.31646, -90.31632, -90.31590, -90.31583, -90.31572, -90.31571, -90.31570, -90.31562, -90.31539, -90.31531, -90.31531, -90.31519, -90.31513, -90.31476, -90.31439, -90.31413, -90.31370, -90.31336, -90.31304, -90.31265, -90.31226, -90.31130, -90.31116, -90.31064, -90.31060, -90.31053, -90.31047, -90.31037, -90.31018, -90.30997, -90.30969, -90.30940, -90.30919, -90.30900, -90.30898, -90.30895, -90.30894, -90.30889, -90.30877, -90.30871, -90.30866, -90.30817, -90.30800, -90.30795, -90.30783, -90.30721, -90.30717, -90.30644, -90.30643, -90.30626, -90.30625, -90.30608, -90.30591, -90.30584, -90.30583, -90.30575, -90.30568, -90.30552, -90.30532, -90.30513, -90.30494, -90.30481, -90.30450, -90.30424, -90.30376, -90.30368, -90.30362, -90.30360, -90.30340, -90.30322, -90.30315, -90.30310, -90.30309, -90.30299, -90.30285, -90.30285, -90.30283, -90.30275, -90.30275, -90.30258, -90.30239, -90.30217, -90.30216, -90.30197, -90.30195, -90.30192, -90.30183, -90.30171, -90.30148, -90.30115, -90.30074, -90.29952, -90.29951, -90.29885, -90.29817, -90.29806, -90.29801, -90.29789, -90.29775, -90.29755, -90.29705, -90.29688, -90.29616, -90.29542, -90.29523, -90.29467, -90.29393, -90.29289, -90.29238, -90.29214, -90.29187, -90.29080, -90.29005, -90.28941, -90.28915, -90.28881, -90.28821, -90.28763, -90.28734, -90.28684, -90.28603, -90.28582, -90.28509, -90.28439, -90.28377, -90.28334, -90.28278, -90.28219, -90.28144, -90.28111, -90.28083, -90.27992, -90.27881, -90.27715, -90.27528, -90.27524, -90.27494, -90.27462, -90.27405, -90.27377, -90.27269, -90.27250, -90.27151, -90.27010, -90.26936, -90.26866, -90.26847, -90.26835, -90.26811, -90.26808, -90.26807, -90.26797, -90.26791, -90.26756, -90.26707, -90.26667, -90.26661, -90.26656, -90.26655, -90.26643, -90.26621, -90.26616, -90.26606, -90.26591, -90.26579, -90.26572, -90.26552, -90.26542, -90.26524, -90.26503, -90.26498, -90.26468, -90.26444, -90.26438, -90.26408, -90.26386, -90.26364, -90.26344, -90.26309, -90.26212, -90.26200, -90.26181, -90.26100, -90.26096, -90.25997, -90.25990, -90.25932, -90.25872, -90.25812, -90.25647, -90.25620, -90.25579, -90.25400, -90.25311, -90.25277, -90.25227, -90.25177, -90.25174, -90.25013, -90.24923, -90.24829, -90.24733, -90.24640, -90.24510, -90.24392, -90.24061, -90.24046, -90.23994, -90.23991, -90.23984, -90.23972, -90.23880, -90.23818, -90.23812, -90.23802, -90.23699, -90.23534, -90.23442, -90.23386, -90.23115, -90.23009, -90.23005, -90.22810, -90.22803, -90.22547, -90.22516, -90.22480, -90.22387, -90.22340, -90.22187, -90.22168, -90.22160, -90.22123, -90.22056, -90.21976, -90.21835, -90.21817, -90.21812, -90.21809, -90.21528, -90.21498, -90.21387, -90.21385, -90.21268, -90.21256, -90.21247, -90.21155, -90.20795, -90.20653, -90.20563, -90.20554, -90.20412, -90.20287, -90.20280, -90.20208, -90.20199, -90.20087, -90.19988, -90.19773, -90.19735, -90.19727, -90.19654, -90.19651, -90.19607, -90.19590, -90.19564, -90.19546, -90.19415, -90.19406, -90.19325, -90.19303, -90.19196, -90.19099, -90.18993, -90.18955, -90.18942, -90.18923, -90.18842, -90.18805, -90.18626, -90.18607, -90.18581, -90.18574, -90.18573, -90.18560, -90.18547, -90.18505, -90.18487, -90.18464, -90.18460, -90.18451, -90.18446, -90.18431, -90.18412, -90.18387, -90.18394, -90.18402, -90.18419, -90.18420, -90.18406, -90.18369, -90.18324, -90.18210, -90.17916, -90.17695, -90.17528, -90.16778, -90.16657, -90.16828, -90.16911, -90.17043, -90.17105, -90.17205, -90.17213, -90.17239, -90.17345, -90.17457, -90.17485, -90.17498, -90.17533, -90.17590, -90.17911, -90.18361, -90.18756, -90.19391, -90.19776, -90.20023, -90.20340, -90.20521, -90.20650, -90.20868, -90.20977, -90.20991, -90.21015, -90.21017, -90.21110, -90.21118, -90.21141, -90.21191, -90.21194, -90.21252, -90.21298, -90.21189, -90.21090, -90.21024, -90.20815, -90.20602, -90.20254, -90.20164, -90.19876, -90.19763, -90.19684, -90.19622, -90.19570, -90.19410, -90.19181, -90.19078, -90.18966, -90.18863, -90.18781, -90.18713, -90.18679, -90.18669, -90.18663, -90.18560, -90.18491, -90.18449, -90.18401, -90.18367, -90.18317, -90.18272, -90.18161, -90.18145, -90.18130, -90.18061, -90.18009, -90.17931, -90.17927, -90.17934, -90.17908, -90.17858, -90.17852, -90.17850, -90.17823, -90.17787, -90.17779, -90.17768, -90.17765, -90.17751, -90.17763, -90.17790, -90.17844, -90.17845, -90.17845, -90.17853, -90.17878, -90.17881, -90.17889, -90.17938, -90.18088, -90.18245, -90.18256, -90.18259, -90.18260, -90.18261, -90.18263, -90.18264, -90.18275, -90.18288, -90.18361, -90.18370, -90.18378, -90.18380, -90.18437, -90.18484, -90.18494, -90.18503, -90.18513, -90.18526, -90.18597, -90.18691, -90.18767, -90.18855, -90.19025, -90.19244, -90.19345, -90.19443, -90.19581, -90.19739, -90.19832, -90.19963, -90.20101, -90.20161, -90.20319, -90.20460, -90.20549, -90.21090, -90.21169, -90.21434, -90.21802, -90.21906, -90.22038, -90.22110, -90.22210, -90.22306, -90.22405, -90.22932, -90.22971, -90.23232, -90.23243, -90.23416, -90.23467, -90.23476, -90.23492, -90.23551, -90.23620, -90.23625, -90.23670, -90.23692, -90.23703, -90.23759, -90.23781, -90.23825, -90.23957, -90.23981, -90.24001, -90.24057, -90.24127, -90.24225, -90.24281, -90.24466, -90.24719, -90.24826, -90.24917, -90.25042, -90.25233, -90.25473, -90.25661, -90.25736, -90.25741, -90.25876, -90.26035, -90.26091, -90.26212, -90.26229, -90.26242, -90.26243, -90.26264, -90.26310, -90.26338, -90.26375, -90.26382, -90.26405, -90.26424, -90.26444, -90.26533, -90.26535, -90.26694, -90.26785, -90.26812, -90.26850, -90.26904, -90.26915, -90.26921, -90.26967, -90.27034, -90.27077, -90.27103, -90.27137, -90.27137, -90.27139, -90.27139, -90.27139, -90.27111, -90.27099, -90.27063, -90.27051, -90.27056, -90.27093, -90.27112, -90.27117, -90.27172, -90.27163, -90.27158, -90.27170, -90.27186, -90.27196, -90.27216, -90.27224, -90.27223, -90.27222, -90.27223, -90.27223, -90.27223, -90.27225, -90.27226, -90.27231, -90.27235, -90.27272, -90.27320, -90.27347, -90.27362, -90.27369, -90.27397, -90.27415, -90.27485, -90.27515, -90.27524, -90.27550, -90.27559, -90.27576, -90.27581, -90.27631, -90.27649, -90.27707, -90.27768, -90.27798, -90.27817, -90.27873, -90.27924, -90.27930, -90.27932, -90.27947, -90.27949, -90.27954, -90.27963, -90.27990, -90.27996, -90.27999, -90.28004, -90.28017, -90.28020, -90.28026, -90.28031, -90.28048, -90.28053, -90.28056, -90.28074, -90.28080, -90.28139, -90.28160, -90.28188, -90.28270, -90.28339, -90.28380, -90.28450, -90.28463, -90.28540, -90.28705, -90.28974, -90.28981, -90.29230, -90.29242, -90.29250, -90.29343, -90.29462, -90.29657, -90.29698, -90.29702, -90.29726, -90.29792, -90.29994, -90.30255, -90.30399, -90.30500, -90.30696, -90.30718, -90.30760, -90.31002, -90.31042, -90.31358, -90.31401, -90.31403, -90.31404, -90.31429, -90.31468, -90.31488, -90.31503, -90.31534, -90.31549, -90.31583, -90.31623, -90.31624, -90.31625, -90.31745, -90.31828, -90.31871, -90.31907, -90.31920, -90.31921, -90.31927, -90.31947, -90.31951, -90.31953, -90.31958, -90.31968, -90.31980, -90.32009, -90.32051, -90.32052, 38.59415, 38.59418, 38.59641, 38.59654, 38.59702, 38.59797, 38.59814, 38.59826, 38.59832, 38.59855, 38.59871, 38.59892, 38.59909, 38.59947, 38.60002, 38.60073, 38.60191, 38.60282, 38.60292, 38.60400, 38.60446, 38.60481, 38.60587, 38.60602, 38.60625, 38.60628, 38.60631, 38.60654, 38.60722, 38.60745, 38.60746, 38.60773, 38.60786, 38.60880, 38.60974, 38.61041, 38.61151, 38.61235, 38.61320, 38.61418, 38.61516, 38.61760, 38.61831, 38.62087, 38.62113, 38.62140, 38.62168, 38.62216, 38.62311, 38.62416, 38.62551, 38.62697, 38.62798, 38.62889, 38.62901, 38.62915, 38.62921, 38.62947, 38.63003, 38.63035, 38.63046, 38.63165, 38.63206, 38.63217, 38.63247, 38.63295, 38.63298, 38.63354, 38.63359, 38.63447, 38.63455, 38.63553, 38.63644, 38.63683, 38.63687, 38.63731, 38.63770, 38.63858, 38.63966, 38.64072, 38.64176, 38.64244, 38.64309, 38.64362, 38.64461, 38.64477, 38.64489, 38.64500, 38.64618, 38.64725, 38.64768, 38.64804, 38.64805, 38.64868, 38.64948, 38.64952, 38.64963, 38.65008, 38.65012, 38.65114, 38.65225, 38.65355, 38.65362, 38.65480, 38.65491, 38.65507, 38.65564, 38.65582, 38.65612, 38.65654, 38.65706, 38.65861, 38.65863, 38.65946, 38.66033, 38.66047, 38.66053, 38.66068, 38.66086, 38.66111, 38.66175, 38.66197, 38.66288, 38.66381, 38.66405, 38.66476, 38.66570, 38.66700, 38.66765, 38.66795, 38.66828, 38.66964, 38.67058, 38.67139, 38.67172, 38.67214, 38.67290, 38.67363, 38.67400, 38.67463, 38.67565, 38.67591, 38.67687, 38.67773, 38.67849, 38.67903, 38.67971, 38.68044, 38.68143, 38.68180, 38.68212, 38.68329, 38.68471, 38.68683, 38.68922, 38.68928, 38.68966, 38.69005, 38.69078, 38.69114, 38.69249, 38.69273, 38.69398, 38.69577, 38.69670, 38.69756, 38.69780, 38.69796, 38.69825, 38.69830, 38.69831, 38.69844, 38.69853, 38.69897, 38.69960, 38.70009, 38.70015, 38.70022, 38.70024, 38.70042, 38.70075, 38.70073, 38.70087, 38.70106, 38.70120, 38.70129, 38.70155, 38.70167, 38.70190, 38.70217, 38.70222, 38.70261, 38.70291, 38.70298, 38.70337, 38.70364, 38.70392, 38.70417, 38.70461, 38.70584, 38.70598, 38.70622, 38.70722, 38.70727, 38.70852, 38.70860, 38.70934, 38.71009, 38.71084, 38.71291, 38.71325, 38.71376, 38.71602, 38.71714, 38.71758, 38.71822, 38.71885, 38.71889, 38.71952, 38.71987, 38.72024, 38.72062, 38.72098, 38.72150, 38.72196, 38.72326, 38.72332, 38.72352, 38.72353, 38.72356, 38.72361, 38.72397, 38.72421, 38.72424, 38.72427, 38.72468, 38.72533, 38.72569, 38.72591, 38.72697, 38.72739, 38.72741, 38.72818, 38.72820, 38.72887, 38.72895, 38.72906, 38.72935, 38.72953, 38.73010, 38.73028, 38.73036, 38.73073, 38.73145, 38.73230, 38.73339, 38.73348, 38.73351, 38.73352, 38.73490, 38.73505, 38.73560, 38.73561, 38.73618, 38.73626, 38.73633, 38.73702, 38.73974, 38.74102, 38.74184, 38.74193, 38.74322, 38.74435, 38.74441, 38.74511, 38.74519, 38.74618, 38.74703, 38.74892, 38.74925, 38.74936, 38.75024, 38.75027, 38.75080, 38.75100, 38.75133, 38.75153, 38.75276, 38.75286, 38.75360, 38.75380, 38.75479, 38.75582, 38.75695, 38.75735, 38.75749, 38.75770, 38.75856, 38.75904, 38.76135, 38.76192, 38.76255, 38.76274, 38.76290, 38.76328, 38.76364, 38.76483, 38.76587, 38.76695, 38.76715, 38.76758, 38.76783, 38.76853, 38.76944, 38.77059, 38.77138, 38.77238, 38.77432, 38.77443, 38.77441, 38.77435, 38.77426, 38.77413, 38.77380, 38.77355, 38.77337, 38.77257, 38.77245, 38.77015, 38.76872, 38.76677, 38.76586, 38.76440, 38.76432, 38.76394, 38.76241, 38.76078, 38.76033, 38.76013, 38.75901, 38.75503, 38.75027, 38.74675, 38.74480, 38.74179, 38.73948, 38.73736, 38.73488, 38.73215, 38.73067, 38.72853, 38.72624, 38.72605, 38.72529, 38.72523, 38.72233, 38.72207, 38.72135, 38.71795, 38.71622, 38.71451, 38.71199, 38.70983, 38.70703, 38.70440, 38.70085, 38.69740, 38.69330, 38.69228, 38.68964, 38.68876, 38.68815, 38.68767, 38.68730, 38.68594, 38.68328, 38.68191, 38.68023, 38.67855, 38.67759, 38.67640, 38.67530, 38.67489, 38.67464, 38.67234, 38.67061, 38.66905, 38.66772, 38.66703, 38.66622, 38.66510, 38.66037, 38.65967, 38.65920, 38.65737, 38.65573, 38.65244, 38.65093, 38.64998, 38.64850, 38.64612, 38.64586, 38.64574, 38.64443, 38.64265, 38.64201, 38.64122, 38.64091, 38.63960, 38.63646, 38.63378, 38.63097, 38.63095, 38.63092, 38.63054, 38.62932, 38.62916, 38.62892, 38.62666, 38.62235, 38.61834, 38.61806, 38.61798, 38.61795, 38.61791, 38.61787, 38.61785, 38.61754, 38.61718, 38.61520, 38.61495, 38.61462, 38.61453, 38.61348, 38.61195, 38.61164, 38.61137, 38.61110, 38.61081, 38.60948, 38.60790, 38.60658, 38.60504, 38.60228, 38.59862, 38.59731, 38.59604, 38.59453, 38.59300, 38.59219, 38.59119, 38.58990, 38.58934, 38.58813, 38.58735, 38.58673, 38.58334, 38.58255, 38.58062, 38.57829, 38.57769, 38.57714, 38.57680, 38.57638, 38.57573, 38.57391, 38.56933, 38.56898, 38.56672, 38.56655, 38.56378, 38.56297, 38.56282, 38.56257, 38.56163, 38.56053, 38.56045, 38.55972, 38.55937, 38.55920, 38.55830, 38.55795, 38.55741, 38.55579, 38.55549, 38.55525, 38.55459, 38.55319, 38.55261, 38.55195, 38.55069, 38.54856, 38.54655, 38.54437, 38.54288, 38.54031, 38.53673, 38.53375, 38.53193, 38.53185, 38.53243, 38.53313, 38.53327, 38.53377, 38.53386, 38.53385, 38.53386, 38.53388, 38.53405, 38.53395, 38.53382, 38.53381, 38.53377, 38.53373, 38.53370, 38.53369, 38.53370, 38.53423, 38.53537, 38.53573, 38.53696, 38.53748, 38.53753, 38.53757, 38.53782, 38.53813, 38.53842, 38.53863, 38.53919, 38.53924, 38.53928, 38.53950, 38.53956, 38.54065, 38.54081, 38.54129, 38.54145, 38.54173, 38.54262, 38.54309, 38.54322, 38.54454, 38.54473, 38.54484, 38.54523, 38.54575, 38.54606, 38.54672, 38.54696, 38.54732, 38.54763, 38.54791, 38.54799, 38.54810, 38.54812, 38.54814, 38.54831, 38.54839, 38.54859, 38.54886, 38.54901, 38.54909, 38.54912, 38.54926, 38.54935, 38.54957, 38.54966, 38.54969, 38.54978, 38.54981, 38.54988, 38.54989, 38.55003, 38.55008, 38.55027, 38.55046, 38.55058, 38.55069, 38.55110, 38.55143, 38.55147, 38.55148, 38.55156, 38.55157, 38.55161, 38.55166, 38.55181, 38.55185, 38.55186, 38.55189, 38.55195, 38.55197, 38.55200, 38.55203, 38.55214, 38.55217, 38.55219, 38.55229, 38.55232, 38.55263, 38.55275, 38.55289, 38.55333, 38.55368, 38.55389, 38.55425, 38.55431, 38.55466, 38.55542, 38.55702, 38.55707, 38.55892, 38.55900, 38.55904, 38.55975, 38.56061, 38.56203, 38.56233, 38.56237, 38.56259, 38.56325, 38.56524, 38.56782, 38.56925, 38.57025, 38.57219, 38.57242, 38.57283, 38.57523, 38.57563, 38.57877, 38.57975, 38.57980, 38.57980, 38.58035, 38.58122, 38.58167, 38.58198, 38.58268, 38.58301, 38.58376, 38.58467, 38.58469, 38.58471, 38.58738, 38.58924, 38.59018, 38.59099, 38.59122, 38.59125, 38.59141, 38.59200, 38.59215, 38.59217, 38.59224, 38.59239, 38.59256, 38.59319, 38.59413, 38.59415