Major Bodies of Water in the City of St. Louis

A simple features data set containing polygons for the major bodies of water in the city.

data(stl_sf_hydro)

Format

A simple feature with 2 rows and 4 variables:

HYDROID

Census identification number

NAME

name of body of water

SQMETER

area covered by water, square meters

geometry

simple features geometry

Examples

class(stl_sf_hydro)

#> [1] "sf"         "data.frame"

str(stl_sf_hydro)

#> Classes ‘sf’ and 'data.frame':	2 obs. of  4 variables:
#>  $ HYDROID : chr  "1102212977721" "110451975812"
#>  $ NAME    : chr  "Mississippi River" "River Des Peres"
#>  $ SQMETER : num  10229167 440873
#>  $ geometry:sfc_POLYGON of length 2; first list element: List of 1
#>   ..$ : num [1:870, 1:2] 888812 888890 888957 889035 889073 ...
#>   ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA
#>   ..- attr(*, "names")= chr  "HYDROID" "NAME" "SQMETER"

head(stl_sf_hydro)

#> Simple feature collection with 2 features and 3 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 871882.3 ymin: 982585.2 xmax: 915266.8 ymax: 1070614
#> epsg (SRID):    NA
#> proj4string:    +proj=tmerc +lat_0=35.83333333333334 +lon_0=-90.5 +k=0.9999333333333333 +x_0=250000 +y_0=0 +datum=NAD83 +units=us-ft +no_defs
#>         HYDROID              NAME  SQMETER                       geometry
#> 1 1102212977721 Mississippi River 10229167 POLYGON ((888811.7 983116.6...
#> 2  110451975812   River Des Peres   440873 POLYGON ((871882.3 1005976,...